Yul: Add toDot to SSACFG
