+++ libyul/CMakeLists.txt
@@ -69,6 +69,7 @@ add_library(yul
 	backends/evm/SSACFGLoopNestingForest.h
 	backends/evm/SSACFGTopologicalSort.cpp
 	backends/evm/SSACFGTopologicalSort.h
+	backends/evm/SSAControlFlowGraph.cpp
 	backends/evm/SSAControlFlowGraph.h
 	backends/evm/SSAControlFlowGraphBuilder.cpp
 	backends/evm/SSAControlFlowGraphBuilder.h
