+++ libyul/backends/evm/SSAControlFlowGraph.h
@@ -206,6 +206,8 @@ public:
 		yulAssert(it->second.value < m_valueInfos.size());
 		return it->second;
 	}
+
+	std::string toDot(bool _includeDiGraphDefinition=true, std::optional<size_t> _functionIndex=std::nullopt) const;
 private:
 	std::deque<ValueInfo> m_valueInfos;
 	std::map<u256, ValueId> m_literals;
